+++ src/main/java/net/dv8tion/jda/internal/entities/EntityBuilder.java
@@ -2504,7 +2504,7 @@ public class EntityBuilder extends AbstractEntityBuilder
             })
             .orElse(Collections.emptyMap());
 
-        final int approxUserInstallCount = object.getInt("approximate_user_install_count", -1);
+        final long approxUserInstallCount = object.getLong("approximate_user_install_count", -1);
 
         return new ApplicationInfoImpl(getJDA(), description, doesBotRequireCodeGrant, iconId, id, flags, isBotPublic, name,
                 termsOfServiceUrl, privacyPolicyUrl, owner, team, tags, redirectUris, interactionsEndpointUrl,
